+++ web/tests/kaf/canvas/test_2310.py
@@ -15,7 +15,7 @@ import ctypes
 class Test:
     #================================================================================================================================
     # @Author: Inbar Willman
-    # Test Name : Canvas: Embed on BSE from upload
+    # Test Name : Canvas: Embed on BSE from upload - v3
     # Test description:
     # Upload entry
     # Go course -> Go to 'Announcements' tab -> Add new announcement -> In announcement, click on wysiwyg and choose media to embed from upload page
@@ -50,30 +50,44 @@ class Test:
             #initialize all the basic vars and start playing
             self,self.driver = clsTestService.initializeAndLoginAsUser(self, driverFix)
             self.common = Common(self.driver)
-            self.entryName = clsTestService.addGuidToString("Embed from upload", self.testNum)
-            self.announcementName = clsTestService.addGuidToString("Embed announcement from upload", self.testNum)
+            self.entryName = clsTestService.addGuidToString("Embed from upload v3", self.testNum)
+            self.announcementName = clsTestService.addGuidToString("Embed announcement from upload v3", self.testNum)
             
             ######################### TEST STEPS - MAIN FLOW #######################
-            writeToLog("INFO","Step 1: Going to to create embed announcement from upload")    
+            if LOCAL_SETTINGS_ENV_NAME != 'ProdNewUI':
+                localSettings.LOCAL_SETTINGS_KMS_ADMIN_URL = 'https://1765561.kaftest.dev.kaltura.com/admin'
+                localSettings.LOCAL_SETTINGS_ADMIN_USERNAME = 'liatv21@mailinator.com'
+                localSettings.LOCAL_SETTINGS_ADMIN_PASSWORD = 'Kaltura1!'
+            else: # testing 
+                localSettings.LOCAL_SETTINGS_KMS_ADMIN_URL = ' https://1665211-10.kaf.kaltura.com/admin'
+                localSettings.LOCAL_SETTINGS_ADMIN_USERNAME = 'liat@mailinator.com'
+                localSettings.LOCAL_SETTINGS_ADMIN_PASSWORD = 'Kaltura1!'
+                
+            writeToLog("INFO","Step 1: Going to set enableNewBSEUI to v3")    
+            if self.common.admin.enableNewBSEUI('v3') == False:
+                writeToLog("INFO","Step 1: FAILED to set enableNewBSEUI to v3")
+                return
+                        
+            writeToLog("INFO","Step 2: Going to to create embed announcement from upload")    
             if self.common.canvas.createEmbedAnnouncements(self.announcementName, self.entryName, embedFrom=Location.UPLOAD_PAGE_EMBED, filePath=self.filePath, description=self.description, isTagsNeeded=False) == False:
                 self.status = "Fail"
-                writeToLog("INFO","Step 1: FAILED to create embed announcement from upload")
+                writeToLog("INFO","Step 2: FAILED to create embed announcement from upload")
                 return  
              
-            writeToLog("INFO","Step 2: Going to to verify embed announcement")    
+            writeToLog("INFO","Step 3: Going to to verify embed announcement")    
             if self.common.kafGeneric.verifyEmbedEntry(self.announcementName, self.uploadThumbnailExpectedResult, '', enums.Application.CANVAS) == False:
                 self.status = "Fail"
-                writeToLog("INFO","Step 2: FAILED to verify embed announcement")
+                writeToLog("INFO","Step 3: FAILED to verify embed announcement")
                 return     
             
-            writeToLog("INFO","Step 3: Going to to delete embed announcement")    
+            writeToLog("INFO","Step 4: Going to to delete embed announcement")    
             if self.common.canvas.deleteAnnouncemnt(self.announcementName) == False:
                 self.status = "Fail"
-                writeToLog("INFO","Step 3: FAILED to delete embed announcement")
+                writeToLog("INFO","Step 4: FAILED to delete embed announcement")
                 return                                          
             
             #########################################################################
-            writeToLog("INFO","TEST PASSED: 'Embed on BSE from upload page' was done successfully")
+            writeToLog("INFO","TEST PASSED: 'Embed on BSE (v3) from upload page' was done successfully")
         # If an exception happened we need to handle it and fail the test       
         except Exception as inst:
             self.status = clsTestService.handleException(self,inst,self.startTime)
